# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectSkill = typeof ProjectSkill[keyof typeof ProjectSkill];


export const ProjectSkill = {
  'ui-design': 'ui-design',
  prototyping: 'prototyping',
  'my-video-edit': 'my-video-edit',
} as const;

export interface Project {
  id: number;
  title: string;
  skill: ProjectSkill;
  description: string;
  year: string;
  /** @nullable */
  assetPath: string | null;
  /** @nullable */
  assetType: string | null;
  /** @nullable */
  assetName: string | null;
  /** @nullable */
  externalUrl: string | null;
  createdAt: string;
}

export type ProjectInputSkill = typeof ProjectInputSkill[keyof typeof ProjectInputSkill];


export const ProjectInputSkill = {
  'ui-design': 'ui-design',
  prototyping: 'prototyping',
  'my-video-edit': 'my-video-edit',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  skill: ProjectInputSkill;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  year: string;
  /** @nullable */
  assetPath?: string | null;
  /** @nullable */
  assetType?: string | null;
  /** @nullable */
  assetName?: string | null;
  /** @nullable */
  externalUrl?: string | null;
}

export type ProjectResponse = Project;

export type ProjectsResponse = Project[];

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @nullable */
  nonce?: string | null;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};