# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List uploaded portfolio projects
      responses:
        "200":
          description: Uploaded portfolio projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectsResponse"
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create an uploaded portfolio project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid project data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{id}:
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete an uploaded portfolio project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Project deleted.
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    Project:
      type: object
      required: [id, title, skill, description, year, assetPath, assetType, assetName, externalUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        skill:
          type: string
          enum: [ui-design, prototyping, my-video-edit]
        description:
          type: string
        year:
          type: string
        assetPath:
          type: ["string", "null"]
        assetType:
          type: ["string", "null"]
        assetName:
          type: ["string", "null"]
        externalUrl:
          type: ["string", "null"]
          format: uri
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title, skill, description, year]
      properties:
        title:
          type: string
          minLength: 1
        skill:
          type: string
          enum: [ui-design, prototyping, my-video-edit]
        description:
          type: string
          minLength: 1
        year:
          type: string
          minLength: 1
        assetPath:
          type: ["string", "null"]
        assetType:
          type: ["string", "null"]
        assetName:
          type: ["string", "null"]
        externalUrl:
          type: ["string", "null"]
          format: uri
    ProjectResponse:
      $ref: "#/components/schemas/Project"
    ProjectsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Project"
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string