# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSkill } from './projectSkill';

export interface Project {
  id: number;
  title: string;
  skill: ProjectSkill;
  description: string;
  year: string;
  /** @nullable */
  assetPath: string | null;
  /** @nullable */
  assetType: string | null;
  /** @nullable */
  assetName: string | null;
  /** @nullable */
  externalUrl: string | null;
  createdAt: Date;
}