# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputSkill } from './projectInputSkill';

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  skill: ProjectInputSkill;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  year: string;
  /** @nullable */
  assetPath?: string | null;
  /** @nullable */
  assetType?: string | null;
  /** @nullable */
  assetName?: string | null;
  /** @nullable */
  externalUrl?: string | null;
}